core: ivshmem: Add support for unidirectional memory regions
Extend the ivshmem device to map up to 3 memory regions. Conceptually,
all regions are optional, thus can be empty but have to be listed as
such in the cell configuration. If present, the first region shall be
read/write for both sides. The second shall be read/write for the local
side and read-only for the remote. The third one shall be read-only
locally and read/writable for the remote. Thus, second and third region
need to be specified reversely in the configuration of both peers.

All three regions are reported via the vendor-specific capability of the
ivshmem device. The capability is extended accordingly.

For the shared-memory network devices, we only use the second and third
region, thus the unidirectional ones. Update the cell configurations
accordingly by splitting the existing one into two halves.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>